+++ enum/NavBar.ts
@@ -1,11 +1,10 @@
 export enum AdminMenuGroup {
-    FAQ_CATEGORY = 'FAQ분류관리',
     FAQ = 'FAQ관리',
     PAGE_INFO = '페이지정보관리',
 }
 
 export enum AdminMenuLabel {
-    FAQ_CATEGORY = '분류관리',
+    FAQ_CATEGORY = 'FAQ분류관리',
     FAQ_LIST = 'FAQ목록',
     FAQ_REGISTER = 'FAQ등록',
     PAGE_INFO = '페이지정보관리',